+++ skill.js
@@ -52,15 +52,18 @@ const adminAccount = JSON.parse(fs.readFileSync(__dirname + '/admin_auth.json'))
 // 스킬 그룹
 router.get('/skill', async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+    
     const connection = await dbPool.getConnection();
 
     const sql = 'select * from skill_group order by touch desc;';
     const [skillGroup] = await connection.query(sql);
+
+    connection.release();
 
     res.render('admin_skillgroup', { skillGroup: skillGroup });
 });
@@ -68,16 +71,19 @@ router.get('/skill', async function (req, res) {
 // 스킬 그룹 등록
 router.post('/skill', async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
     const name = req.body.name;
 
     const connection = await dbPool.getConnection();
     const sql = 'insert into skill_group (name) values (?);';
     await connection.query(sql, [name]);
+
+    connection.release();
 
     res.redirect('/admin/skill');
 
@@ -86,16 +92,19 @@ router.post('/skill', async function (req, res) {
 // 스킬 그룹 터치
 router.get('/skill/:skillGroupId/touch', async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
     const skillGroupId = req.params.skillGroupId;
 
     const connection = await dbPool.getConnection();
     const sql = 'update skill_group set touch = now() where skillGroupId = ?';
     await connection.query(sql, [skillGroupId]);
+
+    connection.release();
 
     res.redirect('/admin/skill');
 });
@@ -103,16 +112,19 @@ router.get('/skill/:skillGroupId/touch', async function (req, res) {
 // 스킬 그룹 삭제
 router.get('/skill/:skillGroupId/delete', async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
     const skillGroupId = req.params.skillGroupId;
 
     const connection = await dbPool.getConnection();
     const sql = 'delete from skill_group where skillGroupId = ?;';
     await connection.query(sql, [skillGroupId]);
+
+    connection.release();
 
     res.redirect('/admin/skill');
 });
@@ -120,16 +132,19 @@ router.get('/skill/:skillGroupId/delete', async function (req, res) {
 // 스킬
 router.get('/skill/:skillGroupId', async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
     const skillGroupId = req.params.skillGroupId;
 
     const connection = await dbPool.getConnection();
     const sql = 'select * from skill where `group` = ? order by touch desc;';
     const [skills] = await connection.query(sql, [skillGroupId]);
+
+    connection.release();
 
     res.render('admin_skill', { skillGroupId: skillGroupId, skills: skills });
 });
@@ -140,11 +155,12 @@ const fileFields = upload.fields([
 ]);
 router.post('/skill/:skillGroupId', fileFields, async function (req, res) {
     // 로그인 확인
-    // const admin = req.session.admin;
-    // if (admin == null) {
-    //     res.redirect('/admin/login');
-    //     return;
-    // }
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
     const skillGroupId = req.params.skillGroupId;
 
     const skillName = req.body.skillName;
@@ -155,9 +171,31 @@ router.post('/skill/:skillGroupId', fileFields, async function (req, res) {
     const sql = 'insert into skill (`group`, skillName, icon) values (?, ?, ?);';
     await connection.query(sql, [skillGroupId, skillName, iconName]);
 
+    connection.release();
+
     res.redirect('/admin/skill/' + skillGroupId);
 });
 
+// 스킬 삭제
+router.get('/skill/:skillGroupId/:skillId/delete', async function (req, res) {
+    // 로그인 확인
+    const admin = req.session.admin;
+    if (admin == null) {
+        res.redirect('/admin/login');
+        return;
+    }
+
+    const skillGroupId = req.params.skillGroupId;
+    const skillId = req.params.skillId;
+
+    const connection = await dbPool.getConnection();
+    const sql = 'delete from skill where skillId = ?;';
+    await connection.query(sql, [skillId]);
+
+    connection.release();
+
+    res.redirect('/admin/skill/' + skillGroupId);
+});
 
 
 module.exports = router;
